requirements.txt
pypdf
mysql-connector-python
python-dotenv
tk
psycopg2
